match load_chunk_4E94B0

+++ Source/map_0x370.hpp
@@ -386,7 +386,7 @@ public:
 
     EXPORT void load_rgen_4E94A0();
 
-    EXPORT void load_chunk_4E94B0(const char_type *pChunkType, s32 chunkLen);
+    EXPORT void load_chunk_4E94B0(const char_type *pChunkType, size_t chunkLen);
 
     EXPORT void process_loaded_zone_data_4E95A0();
